# openapi.yaml
openapi: 3.0.3
info:
  title: MVP Launch API
  description: |
    API for MVP Launch platform - data ingestion, forecasting, and visualization.
  version: 1.0.0
  contact:
    name: Forge Intelligence
    email: dev@forgeintelligence.com
servers:
  - url: https://api.forgeintelligence.com/api/v1
    description: Production server
  - url: http://localhost:8000/api/v1
    description: Local development server
security:
  - bearerAuth: []
...
paths:
...
  /anomalies/scan:
    post:
      tags:
        - Anomaly Detection
      summary: Run anomaly detection scan
      operationId: scanAnomalies
      security:
        - bearerAuth: []
      parameters:
        - name: dataset_id
          in: query
          schema:
            type: string
            format: uuid
        - name: metric_name
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Scan completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  scanned_datasets:
                    type: integer
                  total_anomalies_found:
                    type: integer
                  anomalies_by_dataset:
                    type: object
  /anomalies:
    get:
      tags:
        - Anomaly Detection
      summary: List anomalies
      operationId: listAnomalies
      security:
        - bearerAuth: []
      parameters:
        - name: dataset_id
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            type: string
            enum: [flagged, investigated, dismissed]
        - name: severity
          in: query
          schema:
            type: string
            enum: [low, medium, high]
      responses:
        '200':
          description: List of anomalies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Anomaly'
  /anomalies/{anomaly_id}:
    patch:
      tags:
        - Anomaly Detection
      summary: Update anomaly status
      operationId: updateAnomaly
      security:
        - bearerAuth: []
      parameters:
        - name: anomaly_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  type: string
                  enum: [investigated, dismissed]
                notes:
                  type: string
      responses:
        '200':
          description: Anomaly updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Anomaly'
  /anomalies/notifications:
    get:
      tags:
        - Anomaly Detection
      summary: Get anomaly notifications
      operationId: getNotifications
      security:
        - bearerAuth: []
      parameters:
        - name: unread_only
          in: query
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: List of notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Notification'
  /anomalies/notifications/{notification_id}/read:
    post:
      tags:
        - Anomaly Detection
      summary: Mark notification as read
      operationId: markNotificationRead
      security:
        - bearerAuth: []
      parameters:
        - name: notification_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Notification marked as read
  /anomalies/thresholds:
    post:
      tags:
        - Anomaly Detection
      summary: Set anomaly detection threshold
      operationId: setThreshold
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - metric_name
              properties:
                metric_name:
                  type: string
                z_score_threshold:
                  type: integer
                  default: 3
                  minimum: 1
                  maximum: 10
                iqr_multiplier:
                  type: integer
                  default: 3
                  minimum: 1
                  maximum: 10
                enabled:
                  type: boolean
                  default: true
      responses:
        '201':
          description: Threshold set
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Threshold'
    get:
      tags:
        - Anomaly Detection
      summary: List anomaly thresholds
      operationId: listThresholds
      security:
        - bearerAuth: []
      parameters:
        - name: dataset_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of thresholds
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Threshold'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Anomaly:
      type: object
      properties:
        id:
          type: string
          format: uuid
        dataset_id:
          type: string
          format: uuid
        metric_name:
          type: string
        timestamp:
          type: string
          format: date-time
        expected_value:
          type: string
          nullable: true
        actual_value:
          type: string
        severity:
          type: string
          enum: [low, medium, high]
        detection_method:
          type: string
        status:
          type: string
          enum: [flagged, investigated, dismissed]
        investigated_at:
          type: string
          format: date-time
          nullable: true
        notes:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
    Notification:
      type: object
      properties:
        id:
          type: string
          format: uuid
        anomaly_id:
          type: string
          format: uuid
        read:
          type: boolean
        created_at:
          type: string
          format: date-time
        anomaly:
          $ref: '#/components/schemas/Anomaly'
    Threshold:
      type: object
      properties:
        id:
          type: string
          format: uuid
        dataset_id:
          type: string
          format: uuid
        metric_name:
          type: string
        z_score_threshold:
          type: integer
        iqr_multiplier:
          type: integer
        enabled:
          type: boolean